refactor(global): move AutoReconnect from module to global setting
Relocate AutoReconnect from a Misc module to a default-enabled global
setting (GlobalSettingsAutoReconnect), per review feedback. The feature
is registered under GlobalManager and keeps the same reconnect logic and
delay setting.
